Started making component more independent.
